Choose LGPL
Point here is to make available the compiled library free for whatever use, but don't integrate it in other projects withour releasing their source. The latter is highly unlikely, as it is quite small project, but it may push people using it as for its main purpose - being single point of responsibility when dealing with hooks and not everybody starts doing its own versions and end-users have to deal with more hassle than before.